Refactor graph utils
